move state to ControlledHighTable
